+++ packages/local_auth/local_auth_android/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 1.0.14
 
-* Fixes device credential authentication for API versions less than R.
+* Fixes device credential authentication for API versions before R.
 
 ## 1.0.13